Fix error on MacOS when resolving symlinks
Previously when trying to run a full build on the project, the example
project was failing on the gql target with the following error:

> bazel-out/darwin-opt-exec-2B5CBBC6/bin/gqlgen/generate:
> line 6: realpath: command not found

`realpath` is not available on a Mac prior to Ventura. We can work
around using it to avoid the user having to install extras into their
environment.

This patch replaces the use of `realpath` with the standard `pwd -P`
command that is available across both Linux and MacOS. This
implementation does not support the last segment of the path being a
symlink, but our use case doesn't require that at this time.